Improve render performance
Using the benchmark as a measure, I started changing the render
implementation to see what would improve the performance. Using ivars
instead of the attr_readers was one big source of gains. Merging the
overrides on top of the defaults and iterating once was also a large
source of performance improvements. Finally removing the conditionals
for a single `compact!` at the end added some more performance.

I stopped here because every other idea I had was only making it slower.